+++ notificare-geo/src/main/java/re/notifica/geo/gms/LocationReceiver.kt
@@ -8,9 +8,9 @@ import com.google.android.gms.location.Geofence
 import com.google.android.gms.location.GeofencingEvent
 import com.google.android.gms.location.LocationResult
 import re.notifica.Notificare
-import re.notifica.geo.gms.ktx.geoInternal
 import re.notifica.geo.ktx.INTENT_ACTION_GEOFENCE_TRANSITION
 import re.notifica.geo.ktx.INTENT_ACTION_INTERNAL_LOCATION_UPDATED
+import re.notifica.geo.ktx.geoInternal
 import re.notifica.internal.NotificareLogger
 
 internal class LocationReceiver : BroadcastReceiver() {