+++ notificare-push/src/main/java/re/notifica/push/ktx/Augment.kt
@@ -1,14 +1,25 @@
 package re.notifica.push.ktx
 
+import com.google.firebase.messaging.RemoteMessage
 import re.notifica.Notificare
 import re.notifica.NotificareInternalEventsModule
 import re.notifica.ktx.events
+import re.notifica.push.NotificareInternalPush
 import re.notifica.push.NotificarePush
 import re.notifica.push.internal.NotificarePushImpl
 
 @Suppress("unused")
 public fun Notificare.push(): NotificarePush {
     return NotificarePushImpl
+}
+
+@Suppress("unused")
+public fun NotificarePush.isNotificareNotification(remoteMessage: RemoteMessage): Boolean {
+    return remoteMessage.data["x-sender"] == "notificare"
+}
+
+internal fun Notificare.pushInternal(): NotificareInternalPush {
+    return push() as NotificareInternalPush
 }
 
 internal fun Notificare.eventsInternal(): NotificareInternalEventsModule {